Returned SocketInterfaceHandler(self) to service_manager creation call as the reference to socket_interface_manager. Not sure if intention, but nothing was returned prior on line 69 of service_manager.py
Started work to store some info we will receive from CMD client. Not sure where to store this data yet. We will need to make an array of several callsigns the client will listen for, and add an exception for them within the "is_frame_for_me" method in frame_handler.py.

Before, all instances of socket_interface_manager were None within the actual command handlers, and it's created P2PSessions, as nothing was passed for it. When I passed the SocketInterfaceHandler to the CommandSocket as the socket_interface_manager, it was still invalid to use within P2PSessions because it was referencing self.socket_interface_manager.command_server as a CustomThreadedTCPServer, and not a CommandSocket type, so the command_handler was an invalid reference.
Adding the hasattr seems to make sure the command_handler is accessible. Not very sure about this solution, but I can now access the command_handler methods from within each session using self.socket_interface_manager.command_server.command_handler.

When connecting to another station using a "CONNECT" command in the command socket, the other station will not reply to the P2P_CONNECTION_CONNECT packet unless lines 221/222 are present within the connected_irs method. Not sure if this is intentional, or if this was only partially complete due to a move to use an ARQ connection. For now the connection command still calls P2PConnectionCommand, so need to find out when ARQ comes into play.